adopts read password for windows/unix
